requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
discord.py==2.4.0
ffmpeg==1.4
frozenlist==1.4.1
fuzzywuzzy==0.18.0
idna==3.10
lyricsgenius==3.0.1
multidict==6.1.0
PlexAPI==4.15.16
propcache==0.2.0
pycparser==2.22
PyNaCl==1.5.0
PyYAML==6.0.2
requests==2.32.3
setuptools==75.1.0
soupsieve==2.6
urllib3==2.2.3
wheel==0.44.0
yarl==1.15.2
